requirements.txt
Flask>=0.10.1
gunicorn>=18.0
PyYAML>=3.10
pymongo>=2.6.3
requests>=1.2.3
sh>=1.09
Sphinx>=1.2.1
sphinxcontrib-httpdomain>=1.2.0
termcolor==1.1.0
